Remove DelayJump.
